LayoutTests/platform/ios-simulator/fast/events/touch/page-scaled-touch-gesture-click-expected.txt

CONSOLE MESSAGE: line 84: TypeError: undefined is not a function (evaluating 'eventSender.gestureTap(10, 12)')
FAIL: Timed out waiting for notifyDone to be called
This tests basic single touch gesture generation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
